// src/types/resources/ExplanationOfBenefit.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TPeriod } from '../partials/Period';
import { TDateTime } from '../simpleTypes/DateTime';
import { TExplanationOfBenefitRelated } from '../partials/ExplanationOfBenefitRelated';
import { TExplanationOfBenefitPayee } from '../partials/ExplanationOfBenefitPayee';
import { TExplanationOfBenefitCareTeam } from '../partials/ExplanationOfBenefitCareTeam';
import { TExplanationOfBenefitSupportingInfo } from '../partials/ExplanationOfBenefitSupportingInfo';
import { TExplanationOfBenefitDiagnosis } from '../partials/ExplanationOfBenefitDiagnosis';
import { TExplanationOfBenefitProcedure } from '../partials/ExplanationOfBenefitProcedure';
import { TInt } from '../simpleTypes/Int';
import { TExplanationOfBenefitInsurance } from '../partials/ExplanationOfBenefitInsurance';
import { TExplanationOfBenefitAccident } from '../partials/ExplanationOfBenefitAccident';
import { TExplanationOfBenefitItem } from '../partials/ExplanationOfBenefitItem';
import { TExplanationOfBenefitAddItem } from '../partials/ExplanationOfBenefitAddItem';
import { TExplanationOfBenefitAdjudication } from '../partials/ExplanationOfBenefitAdjudication';
import { TExplanationOfBenefitTotal } from '../partials/ExplanationOfBenefitTotal';
import { TExplanationOfBenefitPayment } from '../partials/ExplanationOfBenefitPayment';
import { TAttachment } from '../partials/Attachment';
import { TExplanationOfBenefitProcessNote } from '../partials/ExplanationOfBenefitProcessNote';
import { TExplanationOfBenefitBenefitBalance } from '../partials/ExplanationOfBenefitBenefitBalance';

export type TExplanationOfBenefit = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  status: string;
  type: TCodeableConcept;
  subType?: TCodeableConcept;
  use: string;
  patient: TReference;
  billablePeriod?: TPeriod;
  created: TDateTime;
  enterer?: TReference;
  insurer: TReference;
  provider: TReference;
  priority?: TCodeableConcept;
  fundsReserveRequested?: TCodeableConcept;
  fundsReserve?: TCodeableConcept;
  related?: TExplanationOfBenefitRelated[];
  prescription?: TReference;
  originalPrescription?: TReference;
  payee?: TExplanationOfBenefitPayee;
  referral?: TReference;
  facility?: TReference;
  claim?: TReference;
  claimResponse?: TReference;
  outcome: string;
  disposition?: string;
  preAuthRef?: string[];
  preAuthRefPeriod?: TPeriod[];
  careTeam?: TExplanationOfBenefitCareTeam[];
  supportingInfo?: TExplanationOfBenefitSupportingInfo[];
  diagnosis?: TExplanationOfBenefitDiagnosis[];
  procedure?: TExplanationOfBenefitProcedure[];
  precedence?: TInt;
  insurance: TExplanationOfBenefitInsurance[];
  accident?: TExplanationOfBenefitAccident;
  item?: TExplanationOfBenefitItem[];
  addItem?: TExplanationOfBenefitAddItem[];
  adjudication?: TExplanationOfBenefitAdjudication[];
  total?: TExplanationOfBenefitTotal[];
  payment?: TExplanationOfBenefitPayment;
  formCode?: TCodeableConcept;
  form?: TAttachment;
  processNote?: TExplanationOfBenefitProcessNote[];
  benefitPeriod?: TPeriod;
  benefitBalance?: TExplanationOfBenefitBenefitBalance[];
};